+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - change logging configuration via properties file
 - test for install script
 - option to install via release archive file
+- option to set local maven repository settings
 
 ### Changed